Rename Telemetry namespace -> Metrics
