+++ test/unittests/Microsoft.ServiceFabric.Diagnostics.Tests/Metrics/NullMeterProviderTest.cs
@@ -5,7 +5,7 @@
 
 using Xunit;
 
-namespace Microsoft.ServiceFabric.Diagnostics.Telemetry
+namespace Microsoft.ServiceFabric.Diagnostics.Metrics
 {
     public abstract class NullMeterProviderTest
     {
